+++ src/main/kotlin/com/msg/gcms/domain/attendance/service/QueryCurrentAttendConditionService.kt
@@ -1,10 +1,8 @@
 package com.msg.gcms.domain.attendance.service
 
 import com.msg.gcms.domain.attendance.presentation.data.dto.SearchScheduleDto
-import com.msg.gcms.domain.attendance.presentation.data.dto.UserAttendanceStatusDto
-import java.time.LocalDate
-import java.time.LocalTime
+import com.msg.gcms.domain.attendance.presentation.data.dto.UserAttendanceStatusListDto
 
 interface QueryCurrentAttendConditionService {
-    fun execute(searchScheduleDto: SearchScheduleDto): List<UserAttendanceStatusDto>
+    fun execute(searchScheduleDto: SearchScheduleDto): List<UserAttendanceStatusListDto>
 }
