+++ src/main/kotlin/com/msg/gcms/domain/attendance/service/impl/QueryCurrentAttendConditionServiceImpl.kt
@@ -10,7 +10,7 @@ import com.msg.gcms.domain.club.domain.repository.ClubRepository
 import com.msg.gcms.domain.club.exception.ClubNotFoundException
 import com.msg.gcms.domain.club.exception.HeadNotSameException
 import com.msg.gcms.domain.user.domain.entity.User
-import com.msg.gcms.domain.attendance.presentation.data.dto.UserAttendanceStatusDto
+import com.msg.gcms.domain.attendance.presentation.data.dto.UserAttendanceStatusListDto
 import com.msg.gcms.domain.auth.domain.Role
 import com.msg.gcms.global.util.UserUtil
 import org.springframework.data.repository.findByIdOrNull
@@ -26,7 +26,7 @@ class QueryCurrentAttendConditionServiceImpl(
     private val attendanceRepository: AttendanceRepository,
     private val attendanceConverter: AttendanceConverter
 ) : QueryCurrentAttendConditionService {
-    override fun execute(searchScheduleDto: SearchScheduleDto): List<UserAttendanceStatusDto> {
+    override fun execute(searchScheduleDto: SearchScheduleDto): List<UserAttendanceStatusListDto> {
         val user: User = userUtil.fetchCurrentUser()
 
         val club = clubRepository.findByIdOrNull(searchScheduleDto.clubId) ?: throw ClubNotFoundException()
